refactor: extract issue mapping helpers from github-labels.ts
