+++ core/state-source/github-labels.ts
@@ -1,5 +1,4 @@
 import { eventStore } from '../event-stream/store.js';
-import { resolveState } from '../state-machine/conflict-resolver.js';
 import type { StateName } from '../state-machine/states.js';
 import { isLegalTransition } from '../state-machine/transitions.js';
 import type {
@@ -13,19 +12,13 @@ import type {
   Subscription,
   WorkItem,
 } from './interface.js';
+import type { GithubIssue, GithubMilestone } from './github-issue-mapper.js';
 import {
-  LABEL_GROUPS,
-  extractLabelValue,
-  parseExec,
-  parseMode,
-  parsePriority,
-  parseSchedule,
-  parseWorkItemType,
-} from './label-parsers.js';
-
-function parseIssueNumber(itemId: string): string {
-  return itemId.match(/#(\d+)$/)?.[1] ?? itemId;
-}
+  mapGithubMilestone,
+  mapIssueToWorkItem,
+  parseIssueNumber,
+} from './github-issue-mapper.js';
+import { LABEL_GROUPS } from './label-parsers.js';
 
 /**
  * Schedule values exposed in the UI. `backlog` and `icebox` are aliases for
@@ -57,109 +50,6 @@ export const SCHEDULE_UI_TO_LABEL: Record<ScheduleUIValue, string> = {
   icebox: 'schedule:later',
   'blocked-by': 'schedule:blocked-by',
 };
-
-interface GithubIssueLabel {
-  name: string;
-}
-
-interface GithubMilestone {
-  number: number;
-  title: string;
-  description: string | null;
-  due_on: string | null;
-  state: 'open' | 'closed';
-  open_issues: number;
-  closed_issues: number;
-}
-
-interface GithubIssue {
-  number: number;
-  title: string;
-  body: string | null;
-  labels: GithubIssueLabel[];
-  milestone: GithubMilestone | null;
-  user: { login: string } | null;
-  created_at: string;
-  pull_request?: unknown;
-}
-
-/**
- * Parse issue body for cross-reference patterns.
- * Handles: #42, owner/repo#42, comma/space separated lists, case-insensitive prefix.
- */
-function parseRefs(body: string, prefix: RegExp): string[] {
-  const refs: string[] = [];
-  for (const line of body.split('\n')) {
-    if (!prefix.test(line)) continue;
-    // Extract all #N or owner/repo#N references in this line.
-    const matches = line.matchAll(/(?:[a-zA-Z0-9_.-]+\/[a-zA-Z0-9_.-]+)?#(\d+)/g);
-    for (const m of matches) {
-      refs.push(m[1]);
-    }
-  }
-  return refs;
-}
-
-function parseDependsOn(body: string): string[] {
-  return parseRefs(body, /depends[\s-]+on\s*:?\s/i);
-}
-
-function parseBlocks(body: string): string[] {
-  return parseRefs(body, /blocks\s*:/i);
-}
-
-function mapIssueToWorkItem(issue: GithubIssue, repoRef: string, ownerLogin: string): WorkItem {
-  const labelNames = issue.labels.map((l) => l.name);
-
-  // State: run full conflict resolver; handles multi-label, archived-wins, zero-label cases.
-  const { state } = resolveState(labelNames);
-
-  // Per ADR 0039 the github-labels mapper is the single ingress for label
-  // values — narrow the strings to typed enums here, then trust the type
-  // downstream. Unknown values fall back to the documented DEFAULT_* in
-  // interface.ts (one place to change a default).
-  const type = parseWorkItemType(extractLabelValue(labelNames, LABEL_GROUPS.TYPE));
-  const priority = parsePriority(extractLabelValue(labelNames, LABEL_GROUPS.PRIORITY));
-  const mode = parseMode(extractLabelValue(labelNames, LABEL_GROUPS.MODE));
-  const schedule: Schedule = parseSchedule(extractLabelValue(labelNames, LABEL_GROUPS.SCHEDULE));
-  const exec = parseExec(extractLabelValue(labelNames, LABEL_GROUPS.EXEC));
-
-  const body = issue.body ?? '';
-
-  return {
-    id: `github:${repoRef}#${issue.number}`,
-    externalId: String(issue.number),
-    repoRef,
-    title: issue.title,
-    body,
-    type,
-    priority,
-    mode,
-    state,
-    authorIsOwner: issue.user?.login === ownerLogin,
-    milestoneId: issue.milestone != null ? String(issue.milestone.number) : undefined,
-    milestoneTitle: issue.milestone?.title,
-    schedule,
-    exec,
-    dependsOn: parseDependsOn(body),
-    blocks: parseBlocks(body),
-    createdAt: new Date(issue.created_at),
-  };
-}
-
-function mapGithubMilestone(m: GithubMilestone): Milestone {
-  return {
-    id: String(m.number),
-    title: m.title,
-    number: m.number,
-    description: m.description ?? undefined,
-    dueOn: m.due_on != null ? new Date(m.due_on) : undefined,
-    isActive: m.state === 'open',
-    state: m.state,
-    openIssues: m.open_issues,
-    closedIssues: m.closed_issues,
-  };
-}
 
 export class GitHubLabelsSource implements StateSource {
   private readonly ownerLogin: string;
